requirements.txt
absl-py==2.3.1
attrs==25.3.0
cffi==2.0.0
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.59.2
jax==0.7.1
jaxlib==0.7.1
kiwisolver==1.4.9
matplotlib==3.10.6
mediapipe==0.10.21
mido==1.3.3
ml_dtypes==0.5.3
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
protobuf==4.25.8
pycparser==2.23
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-rtmidi==1.5.8
scipy==1.16.2
sentencepiece==0.2.1
six==1.17.0
sounddevice==0.5.2
